requirements.txt
fastapi
uvicorn
pdf2image
pytesseract
transformers
sentence-transformers
faiss-cpu
numpy
torch
Pillow
python-multipart
werkzeug
